test: 요금계산 정책별 단위 테스트 클래스 분리
